% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fetch.R
\name{assess}
\alias{assess}
\title{Assess Rmd assignment}
\usage{
assess(block, ext, is_rmd)
}
\arguments{
\item{block}{a list whose elements are code blocks (see \code{\link{tangle}})}

\item{ext}{the file extension}

\item{is_rmd}{whether the file is actually an Rmd}
}
\value{
A list including feedback and other info
}
\description{
Examine the submission and provide basic feedback
}

