% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{describe}
\alias{describe}
\title{Describe assignments in a directory}
\usage{
describe(path)
}
\arguments{
\item{dir}{Path to directory containing moodle assignments}
}
\value{
tibble with information about submissions
}
\description{
Describe assignments in a directory
}
