% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fetch.R
\name{is_rmd}
\alias{is_rmd}
\title{Is file an Rmd?}
\usage{
is_rmd(file)
}
\arguments{
\item{file}{name of file you want to check}

\item{check_yaml}{check whether it has a yaml header?}

\item{check_blocks}{check whether it has at least one code block?}
}
\value{
logical value
}
\description{
Perform some basic tests to see whether the file is an RMarkdown file.
}

