% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{login}
\alias{login}
\title{Log into moodle server}
\usage{
login(moodle_url, user = NULL, pass = NULL)
}
\arguments{
\item{moodle_url}{web address of moodle server}

\item{user}{moodle username for logging in}

\item{pass}{password for logging in}
}
\description{
Log into moodle server
}
\details{
Do this once before using any of the other functions in this package that require access to the server.
}
