% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/post.R
\name{make_report_dir}
\alias{make_report_dir}
\title{Generate directory structure for reports}
\usage{
make_report_dir(path, overwrite = FALSE)
}
\description{
Copy the 'submitted' directory structure to 'reports' subdirectory.
}

