% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/peruse.R
\name{peruse}
\alias{peruse}
\title{Peruse response code blocks}
\usage{
peruse(x, block_id)
}
\arguments{
\item{x:}{tibble with the submission data}

\item{block_id:}{either the number or name of the code block}
}
\description{
Peruse response code blocks in browser
}

