% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/post.R
\name{report}
\alias{report}
\title{Generate report based on feedback}
\usage{
report(feedback, sub_id, path, subdir)
}
\arguments{
\item{feedback}{character string with feedback in RMarkdown format}
}
\value{
path to the report
}
\description{
Generate PDF report based on assessment feedback
}

