% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fetch.R
\name{tangle}
\alias{tangle}
\title{Tangle code from Rmd file}
\usage{
tangle(file)
}
\arguments{
\item{file}{filename}
}
\value{
A list containing file blocks
}
\description{
Extract code blocks from RMarkdown file
}

