% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/visit.R
\name{visit}
\alias{visit}
\title{Visit a submitted Rmd file in the script editor}
\usage{
visit(x, id)
}
\arguments{
\item{x}{: tibble with submission information}

\item{id}{: \code{sub_id} number}
}
\description{
Visit a submitted Rmd file in the script editor
}

